refactor(visual-regions): split label attachment helpers
Separate heading, plain-image, in-region, and table lead-in label attachment heuristics into focused modules while keeping the visual region label facade stable.
